Separated Prius Prime from Prius.
By popular demand, as Primes have different characteristics.  

For Prime-specific mods, I used this...
      ret.steerRatio = 15.00   # unknown end-to-end spec
      tire_stiffness_factor = 0.6371
      ret.mass = 3370 * CV.LB_TO_KG + std_cargo

That was because I realized the effective tire stiffness was a product of the above value, as well as tireStiffnessFront_civic & tireStiffnessRear_civic, so I did some math...

0.4.7.2 Comma values
tireStiffnessFront_civic = 85400*1.43353663=122424.028202
tireStiffnessRear_civic = 90000*1.43353663=129018.2967

0.5.0   Comma values
tireStiffnessFront_civic = 192150*0.7933=152432.595
tireStiffnessRear_civic = 202500*0.7933=160643.25

To determine what the new tire_stiffness_factor is, divide the old effective value by the new tireStiffness*...
122424.028202/192150=0.6371273911111111
129018.2967/202500=0.6371273911111111

In theory, my Prime should drive the same as it did in 0.4.7.2.